move openblas check into fixes
