+++ sklearn/metrics/_pairwise_distances_reduction/_argkmin.pyx.tp
@@ -12,7 +12,8 @@ import warnings
 
 from numbers import Integral
 from scipy.sparse import issparse
-from ...utils import check_array, check_scalar, _in_unstable_openblas_configuration
+from ...utils import check_array, check_scalar
+from ...utils.fixes import _in_unstable_openblas_configuration
 from ...utils.fixes import threadpool_limits
 
 {{for name_suffix in ['64', '32']}}
