+++ sklearn/utils/__init__.py
@@ -24,7 +24,6 @@ from ._param_validation import Integral, Interval, validate_params
 from .class_weight import compute_class_weight, compute_sample_weight
 from .deprecation import deprecated
 from .discovery import all_estimators
-from .fixes import parse_version, threadpool_info
 from .murmurhash import murmurhash3_32
 from .validation import (
     _is_arraylike_not_scalar,
@@ -82,39 +81,6 @@ __all__ = [
 IS_PYPY = platform.python_implementation() == "PyPy"
 _IS_32BIT = 8 * struct.calcsize("P") == 32
 _IS_WASM = platform.machine() in ["wasm32", "wasm64"]
-
-
-def _in_unstable_openblas_configuration():
-    """Return True if in an unstable configuration for OpenBLAS"""
-
-    # Import libraries which might load OpenBLAS.
-    import numpy  # noqa
-    import scipy  # noqa
-
-    modules_info = threadpool_info()
-
-    open_blas_used = any(info["internal_api"] == "openblas" for info in modules_info)
-    if not open_blas_used:
-        return False
-
-    # OpenBLAS 0.3.16 fixed instability for arm64, see:
-    # https://github.com/xianyi/OpenBLAS/blob/1b6db3dbba672b4f8af935bd43a1ff6cff4d20b7/Changelog.txt#L56-L58 # noqa
-    openblas_arm64_stable_version = parse_version("0.3.16")
-    for info in modules_info:
-        if info["internal_api"] != "openblas":
-            continue
-        openblas_version = info.get("version")
-        openblas_architecture = info.get("architecture")
-        if openblas_version is None or openblas_architecture is None:
-            # Cannot be sure that OpenBLAS is good enough. Assume unstable:
-            return True
-        if (
-            openblas_architecture == "neoversen1"
-            and parse_version(openblas_version) < openblas_arm64_stable_version
-        ):
-            # See discussions in https://github.com/numpy/numpy/issues/19411
-            return True
-    return False
 
 
 @validate_params(
